+++ libraries/extractor/src/test/java/androidx/media3/extractor/mp4/FragmentedMp4ExtractorParameterizedTest.java
@@ -89,10 +89,19 @@ public final class FragmentedMp4ExtractorParameterizedTest {
   }
 
   @Test
-  public void sampleSeekableMultipleSidx() throws Exception {
-    assertExtractorBehavior(
-        /* closedCaptionFormats= */ ImmutableList.of(),
-        "media/mp4/sample_fragmented_seekable_multi_sidx.mp4");
+  public void sampleSeekableWithMultipleSidx() throws Exception {
+    String file = "media/mp4/sample_fragmented_seekable_multiple_sidx.mp4";
+    ExtractorAsserts.assertBehavior(
+        () ->
+            new FragmentedMp4Extractor(
+                /* subtitleParserFactory= */ new DefaultSubtitleParserFactory(),
+                /* flags= */ FragmentedMp4Extractor.FLAG_MERGE_FRAGMENTED_SIDX,
+                /* timestampAdjuster= */ null,
+                /* sideloadedTrack= */ null,
+                /* closedCaptionFormats= */ ImmutableList.of(),
+                /* additionalEmsgTrackOutput= */ null),
+        file,
+        simulationConfig);
   }
 
   @Test